Anexo de funciones para luminarias

+++ Practica1/Practica1/Movimiento.h
@@ -15,7 +15,7 @@ public:
 	float getGiroHelice();
 	float giroKitY();
 	float giroKitZ();
-	int horaDia();
+	float horaDia();
 	GLfloat  time();
 	float getDeltaTime();
 	~Movimiento();
